auto-generate the C++ soap code at build time
This code must be generated with the same version of the library
we use to build. This is obviously problematic across OS versions.

Add auto-generation the code in the build process. I tried to
do the same for the onvif.h header but failed. Since it does not
seem to be necessary, I stuck to only the proxy files.